Move Attribution into dir
