+++ context/app/static/js/components/DonorDetail.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import ProvTabs from './Detail/ProvTabs';
 import Summary from './Detail/Summary';
-import Attribution from './Detail/Attribution';
+import Attribution from './Detail/Attribution/Attribution';
 import Protocol from './Detail/Protocol';
 import Metadata from './Detail/Metadata';
 import DetailLayout from './Detail/DetailLayout';
